Changed the welcome screen to be a proper welcome screen and you will only see it once, when the app opens. And the menu is also now on the main screen.
