+++ app/src/main/AndroidManifest.xml
@@ -23,44 +23,43 @@
         <activity
             android:name=".SelectedTaskActivity"
             android:label="selected_task_activity"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".MainMenuActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.c301t3.c301t3app.MainActivity" />
+                android:value="com.c301t3.c301t3app.MainMenuActivity" />
         </activity>
         <activity
             android:name=".BiddingScreen"
             android:label="bidding_screen"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".MainMenuActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.c301t3.c301t3app.MainActivity" />
+                android:value="com.c301t3.c301t3app.MainMenuActivity" />
         </activity>
         <activity
             android:name=".MyTasksActivity"
             android:label="my_tasks"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".MainMenuActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.c301t3.c301t3app.MainActivity" />
+                android:value="com.c301t3.c301t3app.MainMenuActivity" />
         </activity>
         <activity android:name=".SimpleLoginActivity" />
-        <activity android:name=".RegisterActivity"></activity>
+        <activity android:name=".RegisterActivity">
+
+        </activity>
         <activity
             android:name=".MainMenuActivity"
-            android:label="main_menu_activity"
-            android:parentActivityName=".MainActivity">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.c301t3.c301t3app.MainActivity"/>
+            android:label="main_menu_activity">
         </activity>
+
         <activity
             android:name=".NewTaskActivity"
             android:label="new_task_activity.xml"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".MainMenuActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.c301t3.c301t3app.MainActivity"/>
+                android:value="com.c301t3.c301t3app.MainMenuActivity" />
         </activity>
     </application>
