+++ app/src/main/java/com/c301t3/c301t3app/MainActivity.java
@@ -3,11 +3,8 @@ package com.c301t3.c301t3app;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 
@@ -16,63 +13,65 @@ public class MainActivity extends AppCompatActivity {
     private MainActivity activity = this;
     // private final TaskPasser taskPasser = new TaskPasser();
     private TaskPasser taskPasser;
+    private Intent loginIntent;
+    private Intent mainMenuIntent;
     /* Here is a good site with a good tutorial for back button and info sharing between activities.
     * https://google-developer-training.gitbooks.io/android-developer-fundamentals-course-practicals/content/en/Unit%201/21_p_create_and_start_activities.html*/
 
-    /// Menu Start Here-----------------------
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu, menu);
-        return true;
-    }
-
-    /**
-     * @param item
-     * @return
-     */
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        //This is the place to handle all the menu items.
-        int id = item.getItemId();
-
-        switch (id) {
-
-            case R.id.Profile:
-                Toast.makeText(getApplicationContext(), "Profile selected", Toast.LENGTH_SHORT).show();
-
-                break;
-
-            case R.id.SignIN:
-                Toast.makeText(getApplicationContext(), "SignIn selected", Toast.LENGTH_SHORT).show();
-                // go to login activity
-                Intent loginIntent = new Intent(activity, SimpleLoginActivity.class);
-                activity.startActivity(loginIntent);
-
-                break;
-
-            case R.id.Logout:
-                Toast.makeText(getApplicationContext(), "Logout selected", Toast.LENGTH_SHORT).show();
-
-                break;
-
-            case R.id.myTasks:
-                Toast.makeText(getApplicationContext(), "MyTasks selected", Toast.LENGTH_SHORT).show();
-                Intent myTaskIntent = new Intent(activity, MyTasksActivity.class);
-                activity.startActivity(myTaskIntent);
-
-
-                break;
-
-            case R.id.MyBids:
-                Toast.makeText(getApplicationContext(), "MyBids selected", Toast.LENGTH_SHORT).show();
-                break;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-
-
-    //-----------Menu  Stuff ends here-----------//
+    //    /// Menu Start Here-----------------------
+//    @Override
+//    public boolean onCreateOptionsMenu(Menu menu) {
+//        getMenuInflater().inflate(R.menu.menu, menu);
+//        return true;
+//    }
+//
+//    /**
+//     * @param item
+//     * @return
+//     */
+//    @Override
+//    public boolean onOptionsItemSelected(MenuItem item) {
+//        //This is the place to handle all the menu items.
+//        int id = item.getItemId();
+//
+//        switch (id) {
+//
+//            case R.id.Profile:
+//                Toast.makeText(getApplicationContext(), "Profile selected", Toast.LENGTH_SHORT).show();
+//
+//                break;
+//
+//            case R.id.SignIN:
+//                Toast.makeText(getApplicationContext(), "SignIn selected", Toast.LENGTH_SHORT).show();
+//                // go to login activity
+//                Intent loginIntent = new Intent(activity, SimpleLoginActivity.class);
+//                activity.startActivity(loginIntent);
+//
+//                break;
+//
+//            case R.id.Logout:
+//                Toast.makeText(getApplicationContext(), "Logout selected", Toast.LENGTH_SHORT).show();
+//
+//                break;
+//
+//            case R.id.myTasks:
+//                Toast.makeText(getApplicationContext(), "MyTasks selected", Toast.LENGTH_SHORT).show();
+//                Intent myTaskIntent = new Intent(activity, MyTasksActivity.class);
+//                activity.startActivity(myTaskIntent);
+//
+//
+//                break;
+//
+//            case R.id.MyBids:
+//                Toast.makeText(getApplicationContext(), "MyBids selected", Toast.LENGTH_SHORT).show();
+//                break;
+//        }
+//
+//        return super.onOptionsItemSelected(item);
+//    }
+//
+//
+//    //-----------Menu  Stuff ends here-----------//
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -80,15 +79,17 @@ public class MainActivity extends AppCompatActivity {
 
         taskPasser = new TaskPasser();
 
-
-        Button mainButton = findViewById(R.id.button_GoToMyTasks);
+        Button mainButton = findViewById(R.id.button_GoToMainMenu);
+        Button loginButton = findViewById(R.id.loginBtn);
 
         mainButton.setOnClickListener(new View.OnClickListener() {
-
+            /**
+             * @param v
+             */
             public void onClick(View v) {
                 setResult(RESULT_OK);
-
-                Intent mainMenuIntent = new Intent(activity, MainMenuActivity.class);
+                //Toast.makeText(getApplicationContext(), v.toString(),Toast.LENGTH_SHORT).show();
+                mainMenuIntent = new Intent(activity, MainMenuActivity.class);
 
                 ArrayList<Task> dummytasklist = new ArrayList<>();
                 Task task0 = new Task("task0","desc0", TaskStatus.COMPLETED);
@@ -112,6 +113,16 @@ public class MainActivity extends AppCompatActivity {
             }
 
         });
+
+    }
+
+    /**
+     * @param view
+     */
+    public void loginClick(View view) {
+        //Toast.makeText(getApplicationContext(), "Login clicked",Toast.LENGTH_SHORT).show();
+        loginIntent = new Intent(activity, SimpleLoginActivity.class);
+        startActivity(loginIntent);
 
     }
